Fixed bug with MouseButtonEvent and the bug where playing two replays in a row wasn't working. Began investigating the lag problem that affects replays

+++ engine/src/main/java/org/terasology/recording/RecordedEvent.java
@@ -61,7 +61,7 @@ public class RecordedEvent {
 
 
 
-    public double getTimestamp() {
+    public long getTimestamp() {
         return timestamp;
     }
